restructured bookpage folder structure
