--- src/components/Bookview.js
@@ -1,17 +1,0 @@
-import BookTitleBar from "./Bookview/BookTitleBar";
-import BookDesc from "./Bookview/BookDesc";
-import BookFeedback from "./Bookview/BookFeedback";
-import ReadingButton from "./Bookview/ReadingButton";
-import PatronPromote from "./Bookview/PatronPromote"
-import LoginPromote from "./Bookview/LoginPromote"
-import DataPromote from "./Bookview/DataPromote"
-
-export {
-  BookTitleBar,
-  BookDesc,
-  BookFeedback,
-  ReadingButton,
-  PatronPromote,
-  LoginPromote,
-  DataPromote
-}
