+++ src/pages/BookPage/components/BookTagList/BookTagInput/BookTagInput.tsx
@@ -4,10 +4,10 @@ import { makeStyles } from '@material-ui/core/styles';
 import AddCircleIcon from '@material-ui/icons/AddCircle';
 import DoneIcon from '@material-ui/icons/Done';
 import clsx from 'clsx';
-import { useSnackbarContext } from '../../../../contexts/SnackbarContext';
-import { useUser } from '../../../../../lib/user';
+import { useSnackbarContext } from '../../../../../contexts/SnackbarContext';
+import { useUser } from '../../../../../../lib/user';
 import Autocomplete from '@material-ui/lab/Autocomplete';
-import { AutocompleteTag } from '../../../../types/apiTypes';
+import { AutocompleteTag } from '../../../../../types/apiTypes';
 
 const useStyles = makeStyles((theme: Theme) => ({
   chip: {
