+++ src/pages/BookPage/components/BookTagList/BookTagList.tsx
@@ -1,7 +1,7 @@
 import { makeStyles } from '@material-ui/core/styles';
 import { Avatar, Chip, CircularProgress, Grid, Paper } from '@material-ui/core';
 import { useSnackbarContext } from '../../../../contexts/SnackbarContext';
-import BookTagInput from '../BookTagInput/BookTagInput';
+import BookTagInput from './BookTagInput/BookTagInput';
 import { useUser } from '../../../../../lib/user';
 import { JoinedTag } from '../../../../types/common';
 import { useTags } from '../../../../hooks/useTags/useTags';
