+++ src/pages/BookPage/components/index.tsx
@@ -1,4 +1,19 @@
-import BookTagInput from './BookTagInput/BookTagInput';
 import BookTagList from './BookTagList/BookTagList';
+import BookDesc from './BookDesc/BookDesc';
+import BookFeedback from './BookFeedback/BookFeedback';
+import BookTitleBar from './BookTitleBar/BookTitleBar';
+import DataPromote from './DataPromote/DataPromote';
+import LoginPromote from './LoginPromote/LoginPromote';
+import PatronPromote from './PatronPromote/PatronPromote';
+import ReadingButton from './ReadingButton/ReadingButton';
 
-export { BookTagInput, BookTagList };
+export {
+  BookTagList,
+  BookDesc,
+  BookFeedback,
+  BookTitleBar,
+  DataPromote,
+  LoginPromote,
+  PatronPromote,
+  ReadingButton,
+};
